requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
Pillow
numpy
